clean up the goal card code a bit
